% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-yocto-logger.R
\name{read_yocto_logger_csv}
\alias{read_yocto_logger_csv}
\alias{read_yocto_spectral_csv}
\alias{AS7343_metadata}
\title{Read data from YoctoPuce CSV files}
\usage{
read_yocto_logger_csv(
  file,
  cols.pattern = NULL,
  dec = ".",
  sep = ";",
  tz = "UTC",
  ...
)

read_yocto_spectral_csv(
  file,
  cols.pattern = "avg",
  cols.rename = TRUE,
  dec = ".",
  sep = ";",
  tz = "UTC",
  ...
)

AS7343_metadata()
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
    Each row of the table appears as one line of the file.  If it does
    not contain an \emph{absolute} path, the file name is
    \emph{relative} to the current working directory,
    \code{\link{getwd}()}. Tilde-expansion is performed where supported.
    This can be a compressed file (see \code{\link{file}}).

    Alternatively, \code{file} can be a readable text-mode
    \link{connection} (which will be opened for reading if
    necessary, and if so \code{\link{close}}d (and hence destroyed) at
    the end of the function call).  (If \code{\link{stdin}()} is used,
    the prompts for lines may be somewhat confusing.  Terminate input
    with a blank line or an \abbr{EOF} signal, \code{Ctrl-D} on Unix and
    \code{Ctrl-Z} on Windows.  Any pushback on \code{stdin()} will be
    cleared before return.)

    \code{file} can also be a complete URL.  (For the supported URL
    schemes, see the \sQuote{URLs} section of the help for
    \code{\link{url}}.)
  }

\item{cols.pattern}{character A string suitable as argument for
\code{pattern} in a call to \code{grep()}. \code{NULL} or
\code{characte(0)} retain all columns.}

\item{dec}{the character used in the file for decimal points.}

\item{sep}{the field separator character.  Values on each line of the
    file are separated by this character.  If \code{sep = ""} (the
    default for \code{read.table}) the separator is \sQuote{white space},
    that is one or more spaces, tabs, newlines or carriage returns.}

\item{tz}{a character string that specifies which time zone to parse the date with. The string
must be a time zone that is recognized by the user's OS.}

\item{...}{
  Arguments passed on to \code{\link[utils:read.table]{utils::read.table}}
  \describe{
    \item{\code{colClasses}}{character.  A vector of classes to be assumed for
    the columns.  If unnamed, recycled as necessary.  If named, names
    are matched with unspecified values being taken to be \code{NA}.

    Possible values are \code{NA} (the default, when
    \code{\link[utils]{type.convert}} is used), \code{"NULL"} (when the column
    is skipped), one of the atomic vector classes (logical, integer,
    numeric, complex, character, raw), or \code{"factor"}, \code{"Date"}
    or \code{"POSIXct"}.  Otherwise there needs to be an \code{as}
    method (from package \pkg{methods}) for conversion from
    \code{"character"} to the specified formal class.

    Note that \code{colClasses} is specified per column (not per
    variable) and so includes the column of row names (if any).
  }
    \item{\code{nrows}}{integer: the maximum number of rows to read in.  Negative
    and other invalid values are ignored.}
    \item{\code{skip}}{integer: the number of lines of the data file to skip before
    beginning to read data.}
    \item{\code{check.names}}{logical.  If \code{TRUE} then the names of the
    variables in the data frame are checked to ensure that they are
    syntactically valid variable names.  If necessary they are adjusted
    (by \code{\link{make.names}}) so that they are, and also to ensure
    that there are no duplicates.}
    \item{\code{comment.char}}{character: a character vector of length one
    containing a single character or an empty string.  Use \code{""} to
    turn off the interpretation of comments altogether.}
  }}

\item{cols.rename}{logical Flag, if \code{TRUE} use channel names from
sensor IC specification as column names, and if \code{FALSE} keep the
names used in the imported file.}
}
\description{
Functions able to directly read and validate data from .CSV files from
the dataloggers biult into YoctoPuce USB modules.
}
\details{
The dataloggers implemented in different USB modules from
YoctoPuce return .CSV files with a consistent format, that varies only
in the number of data columns and their names. Function
\code{read_yocto_logger()} reads any of these files preserving column
names. The UTC times are converted into \code{POSIXct} values and added
under column \code{time}.

Warnings are issued if something unexpected is encountered in the sequence
of UNIX time stamps in the file.

In modules with multiple channels, the user can enable and disable logging
on a channel by channel basis. Thus, the number of data columns can vary,
making it necessary to match columns by name when replacing some of the
default names by shorter or more informative ones. Currently, the
replacement names are hard coded.
}
\examples{
yocto_meteo.file <-
  system.file("extdata", "yocto-meteo-snm.csv",
            package = "AS7343", mustWork = TRUE)

read_yocto_logger_csv(yocto_meteo.file)
read_yocto_logger_csv(yocto_meteo.file, cols.pattern = "avg")
read_yocto_logger_csv(yocto_meteo.file, cols.pattern = "min|max")
read_yocto_logger_csv(yocto_meteo.file, cols.pattern = "temperature")

yocto_spectral.file <-
  system.file("extdata", "yocto-spectral-LED.csv",
            package = "AS7343", mustWork = TRUE)

read_yocto_spectral_csv(yocto_spectral.file)
read_yocto_spectral_csv(yocto_spectral.file, cols.rename = FALSE)
read_yocto_spectral_csv(yocto_spectral.file, cols.pattern = "Channel1\\\\.")

AS7343_metadata()

}
