% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul-response.R
\name{simul_response}
\alias{simul_response}
\alias{simul_AS7343}
\title{Simulate spectral sensor response}
\usage{
simul_response(
  source.spct,
  sensor.mspct,
  norm = "skip",
  range = NULL,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  time.unit = NULL,
  scale.factor = 1
)

simul_AS7343(
  source.spct,
  range = NULL,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  time.unit = NULL,
  scale.factor = 1
)
}
\arguments{
\item{source.spct}{source_spct object Light source spectral irradiance.}

\item{sensor.mspct}{responce_spct or response_mspct object Light sensor
spectral responsiveness, one or more channels.}

\item{norm}{numeric Normalization wavelength (nm) or character string "max",
or "min" for normalization at the corresponding wavelength, "update" to
update the normalization after modifying units of expression, quantity
or range but respecting the previously used criterion, "undo" to revert
an existing normalization or "skip" to force return of \code{x} unchanged.}

\item{range}{numeric vector of length two Range of wavelengths to use
(nanometres, nm)}

\item{unit.out}{character Allowed values \code{"energy"}, and
\code{"photon"}, or its alias \code{"quantum"}.}

\item{time.unit}{character or lubridate::duration object.}

\item{scale.factor}{numeric vector of length 1, or length equal to that of
\code{w.band}. Numeric multiplier applied to returned values.}
}
\description{
Simulate spectral sensor response
}
\details{
Compute sensor response spectrum by convolution of light source
emission spectrum or illumination spectrum and the responsiveness spectrum
of a sensor with one or more channels. Return the integral over wavelengths
for each sensor channel.
}
\examples{
library(photobiologySensors)

simul_response(sun.spct, sensors.mspct$ams_AS7343, norm = "undo")
simul_response(sun.spct, ccd.spct)
simul_response(sun.spct, ccd.spct, range = c(400, 700))
simul_response(sun.spct, ccd.spct, unit.out = "photon")

simul_AS7343(sun.spct, unit.out = "photon")

}
