% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yocto-spectral2mspct.R
\docType{data}
\name{yocto_spectral.df}
\alias{yocto_spectral.df}
\title{Example of imported spectral data}
\format{
A \code{data.frame} with 199 rows and 14 variables.
}
\usage{
yocto_spectral.df
}
\description{
A dataset containing data logged with a Yocto-Spectral USB
module from YoctoPuce and imported with function
\code{\link[rYoctoPuceInOut:read_yocto_logger_csv]{read_yocto_spectral_csv}()}.
}
\examples{
head(yocto_spectral.df)

}
\seealso{
\code{\link{yocto_spectral2mspct}()},
\code{{read_yocto_logger_csv}()}.
}
\keyword{datasets}
