% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yocto-spectral2mspct.R
\name{yocto_spectral2mspct}
\alias{yocto_spectral2mspct}
\title{Yocto-Spectral to spectra}
\usage{
yocto_spectral2mspct(x, channels = "all")
}
\arguments{
\item{x}{data.frame With only one of average, minimum or maximum sensor
counts.}
}
\description{
Convert a data frame with data imported with \code{read_yocto_spectral_csv()}
into a colelction of \code{raw_spct} objects-
}
