% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-yocto-logger.R
\name{read_yocto_logger_csv}
\alias{read_yocto_logger_csv}
\alias{read_yocto_spectral_csv}
\alias{AS7343_metadata}
\title{Read data from YoctoPuce CSV files}
\usage{
read_yocto_logger_csv(
  file,
  geocode = NULL,
  label = NULL,
  settings.file = NULL,
  cols.pattern = NULL,
  cols.logical.names = FALSE,
  nacols.rm = is.null(cols.pattern),
  dec = ".",
  sep = ";",
  tz = "UTC",
  ...
)

read_yocto_spectral_csv(
  file,
  settings.file = NULL,
  cols.pattern = "avg",
  cols.rename = TRUE,
  dec = ".",
  sep = ";",
  tz = "UTC",
  ...
)

AS7343_metadata()
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
    Each row of the table appears as one line of the file.  If it does
    not contain an \emph{absolute} path, the file name is
    \emph{relative} to the current working directory,
    \code{\link{getwd}()}. Tilde-expansion is performed where supported.
    This can be a compressed file (see \code{\link{file}}).

    Alternatively, \code{file} can be a readable text-mode
    \link{connection} (which will be opened for reading if
    necessary, and if so \code{\link{close}}d (and hence destroyed) at
    the end of the function call).  (If \code{\link{stdin}()} is used,
    the prompts for lines may be somewhat confusing.  Terminate input
    with a blank line or an \abbr{EOF} signal, \code{Ctrl-D} on Unix and
    \code{Ctrl-Z} on Windows.  Any pushback on \code{stdin()} will be
    cleared before return.)

    \code{file} can also be a complete URL.  (For the supported URL
    schemes, see the \sQuote{URLs} section of the help for
    \code{\link{url}}.)
  }

\item{geocode}{data.frame Containing columns \code{lon} and \code{lat} used
to set attribute \code{"where.measured"}.}

\item{label}{character Additional text to be appended to the default
value used to set attribute \code{"comment"}.}

\item{settings.file}{character Path to a JSON file containing the module
settings used to acquire the data.}

\item{cols.pattern}{character A string suitable as argument for
\code{pattern} in a call to \code{grep()}. \code{NULL} or
\code{characte(0)} retain all data columns, while \code{NA} returns
all data and time columns.}

\item{cols.logical.names}{logical Use logical names from module
metadata instead of column heading in CSV file. Require readable
\code{settings.file}.}

\item{nacols.rm}{logical If \code{TRUE} delete columns that contain only
\code{NA} values.}

\item{dec}{the character used in the file for decimal points.}

\item{sep}{the field separator character.  Values on each line of the
    file are separated by this character.  If \code{sep = ""} (the
    default for \code{read.table}) the separator is \sQuote{white space},
    that is one or more spaces, tabs, newlines or carriage returns.}

\item{tz}{a character string that specifies which time zone to parse the date with. The string
must be a time zone that is recognized by the user's OS.}

\item{...}{
  Arguments passed on to \code{\link[utils:read.table]{utils::read.table}}
  \describe{
    \item{\code{colClasses}}{character.  A vector of classes to be assumed for
    the columns.  If unnamed, recycled as necessary.  If named, names
    are matched with unspecified values being taken to be \code{NA}.

    Possible values are \code{NA} (the default, when
    \code{\link[utils]{type.convert}} is used), \code{"NULL"} (when the column
    is skipped), one of the atomic vector classes (logical, integer,
    numeric, complex, character, raw), or \code{"factor"}, \code{"Date"}
    or \code{"POSIXct"}.  Otherwise there needs to be an \code{as}
    method (from package \pkg{methods}) for conversion from
    \code{"character"} to the specified formal class.

    Note that \code{colClasses} is specified per column (not per
    variable) and so includes the column of row names (if any).
  }
    \item{\code{nrows}}{integer: the maximum number of rows to read in.  Negative
    and other invalid values are ignored.}
    \item{\code{skip}}{integer: the number of lines of the data file to skip before
    beginning to read data.}
    \item{\code{check.names}}{logical.  If \code{TRUE} then the names of the
    variables in the data frame are checked to ensure that they are
    syntactically valid variable names.  If necessary they are adjusted
    (by \code{\link{make.names}}) so that they are, and also to ensure
    that there are no duplicates.}
    \item{\code{comment.char}}{character: a character vector of length one
    containing a single character or an empty string.  Use \code{""} to
    turn off the interpretation of comments altogether.}
  }}

\item{cols.rename}{logical Flag, if \code{TRUE} use channel names from
sensor IC specification as column names, and if \code{FALSE} keep the
names used in the imported file.}
}
\value{
A data frame with \code{POSIXct} time stamps in column \code{time}
and data in a variable number of columns containing values converted by
\code{\link[utils:read.table]{read.csv}()}.
}
\description{
Functions able to directly read and validate data from .CSV files created by
the data loggers built into YoctoPuce USB modules.
}
\details{
The dataloggers implemented in different USB modules from
YoctoPuce return .CSV files with a consistent format, that varies only
in the number of data columns and their names. Function
\code{read_yocto_logger()} reads any of these files preserving column
names. The UTC times are converted into \code{POSIXct} values and added
under column \code{time}. A subset of columns can be requested by
passing a regular expression to be used in a call to \code{grep()} on the
column names. If \code{nacols.rm = TRUE} is passed, columns containing only
\code{NA} values are deleted, as these columns in most cases correspond
to logger channels that are not in use.

Warnings are issued if something unexpected, such as dates several years
into the past, unsorted or repeated time stamps
are encountered in the sequence of UNIX time stamps in the file. These
discrepancies can occur when the modules' or the YoctoHub clocks do not
acquire a valid time at start-up. Inconsistent time steps are reported
through a message, as these are the result of a stop followed by restart
of logging. During such a pause it is possible that module settings could
have been changed.

If the name of a JSON file as saved from the YoctoPuce module is passed as
an argument to parameter \code{settings.file} its contents parse into an
R list will be saved to attribute \code{yoctopuce.settings} in the returned
data frame.

Function \code{read_yocto_spectral_csv()} is a wrapper on
\code{read_yocto_logger()} that renames columns using the names used in
the data sheet for the AS7343 sensor.

In modules with multiple channels, the user can enable and disable logging
on a channel by channel basis. Thus, the number of data columns can vary,
making it necessary to match columns by name when replacing some of the
default names by shorter or more informative ones.
}
\section{Warning!}{
 The units and basis of expression, and in several cases
even the physical quantity measured cannot be decoded from the CSV files.
The values in the returned data frame are those read from the file, and
the read values have to be interpreted based on the settings and, possibly,
scripts used during data acquisition by the USB module. These settings
if available in a JSON file downloaded from the module can be stored in
attribute \code{yoctopuce.settings} of the data frame returned.
}

\examples{
yocto_meteo.file <-
  system.file("extdata", "yocto-meteo-snm.csv",
            package = "rYoctoPuceInOut", mustWork = TRUE)

read_yocto_logger_csv(yocto_meteo.file)
read_yocto_logger_csv(yocto_meteo.file, cols.pattern = "avg")
read_yocto_logger_csv(yocto_meteo.file, cols.pattern = "min|max")
read_yocto_logger_csv(yocto_meteo.file, cols.pattern = "temperature")
read_yocto_logger_csv(yocto_meteo.file, nrows = 5L)

yocto_mv.file <-
  system.file("extdata", "yocto-millivolt-Rx.csv",
            package = "rYoctoPuceInOut", mustWork = TRUE)

read_yocto_logger_csv(yocto_mv.file)

yocto_v.file <-
  system.file("extdata", "yocto-0-10V-Rx.csv",
            package = "rYoctoPuceInOut", mustWork = TRUE)

read_yocto_logger_csv(yocto_v.file)
read_yocto_logger_csv(yocto_v.file, cols.pattern = "avg")
read_yocto_logger_csv(yocto_v.file, cols.pattern = "min|max")
read_yocto_logger_csv(yocto_v.file, cols.pattern = "Sensor2")

yocto_serial.file <-
  system.file("extdata", "yocto-serial.csv",
            package = "rYoctoPuceInOut", mustWork = TRUE)
yocto_serial_settings.file <-
  system.file("extdata", "YSERIAL1-EAD24.json",
            package = "rYoctoPuceInOut", mustWork = TRUE)

read_yocto_logger_csv(yocto_serial.file)
read_yocto_logger_csv(yocto_serial.file, nacols.rm = FALSE)
read_yocto_logger_csv(yocto_serial.file, cols.pattern = "avg")
read_yocto_logger_csv(yocto_serial.file, cols.pattern = "min|max")
read_yocto_logger_csv(yocto_serial.file, cols.pattern = "Sensor1$")
data.tb <-
  read_yocto_logger_csv(yocto_serial.file, yocto_serial_settings.file)
how_measured(data.tb)
colnames(data.tb)

yocto_i2c.file <-
  system.file("extdata", "yocto-i2c-tsl2591.csv",
            package = "rYoctoPuceInOut", mustWork = TRUE)

read_yocto_logger_csv(yocto_i2c.file)
read_yocto_logger_csv(yocto_i2c.file, cols.pattern = "avg")
read_yocto_logger_csv(yocto_i2c.file, cols.pattern = "min|max")
read_yocto_logger_csv(yocto_i2c.file, cols.pattern = "Sensor1[.]")
read_yocto_logger_csv(yocto_i2c.file, cols.pattern = "Sensor[1-3][.]")
read_yocto_logger_csv(yocto_i2c.file, cols.pattern = "avg", nacols.rm = TRUE)
read_yocto_logger_csv(yocto_i2c.file, cols.pattern = "Sensor[1-3][.]avg")

yocto_CO2.file <-
  system.file("extdata", "yocto-CO2-V1.csv",
            package = "rYoctoPuceInOut", mustWork = TRUE)

read_yocto_logger_csv(yocto_CO2.file)
read_yocto_logger_csv(yocto_CO2.file, cols.pattern = "avg")
read_yocto_logger_csv(yocto_CO2.file, cols.pattern = "min|max")

yocto_spectral.file <-
  system.file("extdata", "yocto-spectral-LED.csv",
            package = "rYoctoPuceInOut", mustWork = TRUE)

read_yocto_spectral_csv(yocto_spectral.file)
read_yocto_spectral_csv(yocto_spectral.file, cols.rename = FALSE)
read_yocto_spectral_csv(yocto_spectral.file, cols.pattern = "Channel1\\\\.")

AS7343_metadata()

}
\references{
Documentation for each YocotoPuce USB module is available at
\url{https://www.yoctopuce.com/}.
}
