% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yocto-spectral2mspct.R
\name{yocto_spectral2mspct}
\alias{yocto_spectral2mspct}
\title{Yocto-Spectral to spectra}
\usage{
yocto_spectral2mspct(df, channels = "all")
}
\arguments{
\item{df}{data.frame With only one of average, minimum or maximum sensor
counts for the channels.}

\item{channels}{character vector of length one or longer. In addition to a
vector of spectral sensor channel names as used in the AS7343
documentation, the nicknames \code{"all"}, \code{"wide"}, \code{"narrow"}
and \code{"xyz"} as vectors of length one are accepted.}
}
\description{
Convert a data frame with data imported with \code{read_yocto_spectral_csv()}
into a colelction of \code{raw_spct} objects-
}
